+++ gems/aws-sdk-s3/CHANGELOG.md
@@ -1,6 +1,8 @@
 Unreleased Changes
 ------------------
 
+* Feature - Fix bug where thread_count option was not being respected for multipart uploads.
+
 1.146.0 (2024-03-18)
 ------------------